+++ modules/cirrus/workflow/inputs.tf
@@ -37,62 +37,101 @@ variable "workflow_config" {
         task. This is necessary for granting the Workflow execution role invoke
         permissions on these functions.
 
-      - template_filepath: (required, string) Path to an Amazon State Machine
-        definition template file. The path must be relative to the ROOT module
-        of the Terraform deployment. The template should use valid Amazon States
-        Language syntax; wherever a Cirrus Task resource ARN is needed, a
-        Terraform interpolation sequence (a "$\{...}" without the "\") may be
-        used instead. The variable name does not matter so long as there is a
-        corresponding entry in the "template_variables" argument.
-        Example template snippet:
+      - default_template_config: (optional, object) Used to create an Amazon
+        State Machine by rendering and combining predefined state configurations
+        using outputs from the Cirrus Tasks referenced in 'state_sequence'. If
+        your State Machine is a simple sequential execution of Cirrus Tasks,
+        this config should be preferred over 'custom_template_config' as it
+        removes the need for creating a State Machine definition template.
+        Contents:
+          - description: (required, string) Workflow description
+          - allow_retry: (required, bool) Whether individual states should be
+            retried in the event of a failure.
+          - state_sequence: (required, list[object]) List of Cirrus Tasks that
+            will be executed sequentially.
+            Contents:
+              - task_name: (required, string) Cirrus Task name. Used for getting
+                Task output resource ARNs. Will also be used for the state name
+                if 'state_name' is omitted.
+              - task_type: (required, string) lambda or batch
+              - state_name: (optional, string) State name. Must be unique within
+                the Workflow. In most cases, this can be omitted as 'task_name'
+                is used by default. Useful if you need to call the same Task
+                multiple times within a Workflow.
 
-          "States": {
-            "FirstState": {
-              "Type": "Task",
-              "Resource": "$\{my-task-lambda}",  // REMOVE THE "\"
-              "Next": "SecondState",
-              ...
-            },
+      - custom_template_config: (optional, object) Used to create an Amazon
+        State Machine by rendering a user-provided template with variables that
+        reference Cirrus Task outputs. Useful for complex State Machines that
+        cannot be created using the 'default_template_config' option.
+        Contents:
+          - filepath: (required, string) Path to an Amazon State Machine
+            template definition. The path must be relative to the ROOT module of
+            the Terraform deployment. The template should use valid Amazon State
+            Language syntax; wherever a Cirrus Task resource ARN is needed, a
+            Terraform interpolation sequence (a "$\{...}" without the "\") may
+            be used instead. The variable name does not matter so long as there
+            is a corresponding entry in the "template_variables" argument.
+            Example template snippet:
 
-        Cirrus may deploy and manage several builtin tasks. Resource ARNs for
-        these tasks may be referenced in a Workflow template using a predefined
-        variable name without having to supply a 'template_variable' entry.
-          - If Batch Tasks were created, the following variables may be used:
-            - PRE-BATCH: cirrus-geo pre-batch Lambda function ARN
-            - POST-BATCH: cirrus-geo post-batch Lambda function ARN
+              "States": {
+                "FirstState": {
+                  "Type": "Task",
+                  "Resource": "$\{my-task-lambda}",  // REMOVE THE "\"
+                  "Next": "SecondState",
+                  ...
+                },
 
-      - template_variables: (optional, map[object]) A map of template variable
-        names to their corresponding Cirrus Task attributes. Assuming a Cirrus
-        Task named "my-task" with Lambda config was passed to the 'task' module,
-        the following workflow template variable config:
+            Cirrus may deploy and manage several builtin tasks. Resource ARNs
+            for these tasks may be referenced in a Workflow template using
+            predefined variable names without having to supply a
+            'template_variable' entry.
+              - If Batch Tasks were created, these variables may be used:
+                - PRE-BATCH: cirrus-geo pre-batch Lambda function ARN
+                - POST-BATCH: cirrus-geo post-batch Lambda function ARN
 
-          my-task-lambda = {
-            task_name = "my-task"
-            task_type = "lambda"
-            task_attr = "function_arn"
-          }
+          - template_variables: (optional, map[object]) Map of template variable
+            names to their corresponding Cirrus Task attributes. Assuming a
+            Cirrus Task named "my-task" with a Lambda config was passed to the
+            'task' module, the following workflow template variable config:
 
-        when used with the example Workflow snippet above would result in the
-        following content after template interpolation:
+              my-task-lambda = {
+                task_name = "my-task"
+                task_type = "lambda"
+                task_attr = "function_arn"
+              }
 
-          "States": {
-            "FirstState": {
-              "Type": "Task",
-              "Resource": "arn:aws:lambda:us-west-2:123456789012:function:my-function",
-              "Next": "SecondState",
-              ...
-            },
+            when used with the example Workflow snippet above, would result in
+            the following content after template interpolation:
+
+              "States": {
+                "FirstState": {
+                  "Type": "Task",
+                  "Resource": "arn:aws:lambda:us-west-2:123456789012:function:my-function",
+                  "Next": "SecondState",
+                  ...
+                },
   DESCRIPTION
 
   type = object({
     name                   = string
     non_cirrus_lambda_arns = optional(list(string))
-    template_filepath      = string
-    template_variables = optional(map(object({
-      task_name = string
-      task_type = string
-      task_attr = string
-    })))
+    default_template_config = optional(object({
+      description = string
+      allow_retry = bool
+      state_sequence = list(object({
+        state_name = optional(string)
+        task_name  = string
+        task_type  = string
+      }))
+    }))
+    custom_template_config = optional(object({
+      filepath = string
+      variables = optional(map(object({
+        task_name = string
+        task_type = string
+        task_attr = string
+      })))
+    }))
   })
 
   # Value must be provided else this module serves no purpose
@@ -100,9 +139,9 @@ variable "workflow_config" {
 
   validation {
     condition = (
-      var.workflow_config.template_variables != null
+      try(var.workflow_config.custom_template_config.variables, null) != null
       ? alltrue([
-        for _, tpl_variable in var.workflow_config.template_variables :
+        for _, tpl_variable in var.workflow_config.custom_template_config.variables :
         (
           contains(
             ["lambda", "batch"],
@@ -118,9 +157,58 @@ variable "workflow_config" {
     )
 
     error_message = <<-ERROR
-      Invalid template variable config. Each key must have a valid value:
+      Invalid variable config provided. Each key must have a valid value:
         - task_type => one of ["lambda", "batch"]
         - task_attr => one of ["function_arn", "job_definition_arn", "job_queue_arn"]
+    ERROR
+  }
+
+  validation {
+    condition = (
+      var.workflow_config.default_template_config != null
+      ? length(distinct([
+        for state in var.workflow_config.default_template_config.state_sequence :
+        coalesce(state.state_name, state.task_name)
+      ])) == length(var.workflow_config.default_template_config.state_sequence)
+      : true
+    )
+    error_message = <<-ERROR
+      Invalid state sequence provided. Each state must have a unique name (the
+      'task_name' attribute is used if 'state_name' was not provided). Set the
+      'state_name' attribute on states as needed to avoid name collisions.
+    ERROR
+  }
+
+  validation {
+    condition = (
+      var.workflow_config.default_template_config != null
+      ? alltrue([
+        for state in var.workflow_config.default_template_config.state_sequence :
+        contains(["lambda", "batch"], state.task_type)
+      ])
+      : true
+    )
+    error_message = <<-ERROR
+      Invalid state sequence provided. Each key must have a valid value:
+        - task_type => one of ["lambda", "batch"]
+    ERROR
+  }
+
+  validation {
+    condition = (
+      (
+        var.workflow_config.default_template_config == null
+        && var.workflow_config.custom_template_config != null
+      )
+      ||
+      (
+        var.workflow_config.default_template_config != null
+        && var.workflow_config.custom_template_config == null
+      )
+    )
+    error_message = <<-ERROR
+      Invalid workflow config provided. Exactly one of 'default_template_config'
+      and 'custom_template_config' must be set.
     ERROR
   }
 }